\name{character or NULL-class}
\docType{class}
\alias{character or NULL-class}
\title{Class "character or NULL"}
\description{A virtual class needed by the "matrix.csc.hb" class}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}


\section{Methods}{
No methods defined with class "character or NULL" in the signature.
}
\keyword{classes}

