\name{mslm-class}
\docType{class}
\alias{mslm-class}
\title{Class "mslm"}
\description{A sparse extension of \code{lm}}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("mslm", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{coefficients}:}{Object of class \code{numeric} estimated coefficients}
    \item{\code{chol}:}{Object of class \code{matrix.csr.chol} generated by 
the function \code{chol}}
    \item{\code{residuals}:}{Object of class \code{"numeric"} residuals}
    \item{\code{fitted}:}{Object of class \code{"numeric"} fitted values}
  }
}
\section{Extends}{
Class \code{"lm"}, directly.
Class \code{"slm"}, directly.
Class \code{"oldClass"}, by class "lm".
}
\section{Methods}{
  \describe{
    \item{coef}{\code{signature(object = "mslm")}: ... }
    \item{fitted}{\code{signature(object = "mslm")}: ... }
    \item{residuals}{\code{signature(object = "mslm")}: ... }
    \item{summary}{\code{signature(object = "mslm")}: ... }
  }
}
\seealso{ \code{\link{slm}}}
\keyword{classes}

