\name{numeric or NULL-class}
\docType{class}
\alias{numeric or NULL-class}
\title{Class "numeric or NULL"}
\description{A virtual class needed by the "matrix.csc.hb" class}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}


\section{Methods}{
No methods defined with class "numeric or NULL" in the signature.
}
\keyword{classes}

