\name{summary.mslm-class}
\docType{class}
\alias{summary.mslm-class}
\title{Class "summary.mslm" }
\description{Sparse version of \code{summary.lm}}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}


\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "summary.mslm")}: ... }
  }
}
\keyword{classes}
