\name{summary.slm-class}
\docType{class}
\alias{summary.slm-class}
\title{Class "summary.slm" }
\description{Sparse version of \code{summary.lm}}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}


\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "summary.slm")}: ... }
  }
}
\keyword{classes}
