\name{triogramX}
\alias{triogramX}
\alias{X}
\title{A Design Matrix for a Triogram Problem}
\description{This is a design matrix arising from a bivariate smoothing
problem using penalized triogram fitting.  It is used in the SparseM vignette
to illustrate the use of the sparse matrix image function.}
\usage{data(triogramX)}
\format{A 375 by 100 matrix stored in compressed sparse row format}
\references{
Koenker, R and Ng, P. (2002).  SparseM:  A Sparse Matrix Package for \R,\cr
\url{http://www.econ.uiuc.edu/~roger/research}
}
\keyword{datasets}
