\name{SparseM.image}
\alias{image.matrix.csr}
\alias{image.matrix.csc}
\alias{SparseM.image}
\title{Image Plot for Sparse Matrices}
\description{Display the pattern of non-zero entries of 
a matrix of class \code{matrix.csr} or \code{matrix.csc}}
\usage{
\method{image}{matrix.csr}(x, col=c("white","gray"),xlab="column",ylab="row", \dots)
}
\arguments{
\item{x}{a matrix of class \code{matrix.csr} or \code{matrix.csc}.}
\item{col}{a list of colors such as that generated by `rainbow'.
          Defaults to c("white","gray")}
\item{xlab,ylab}{ each a character string giving the labels for the x and y
          axis.}
\item{\dots}{additional arguments.}
}
\details{The pattern of the non-zero entries of a sparse matrix 
is displayed.  By default nonzero entries of the matrix appear
as gray blocks and zero entries as white background.}
\references{
Koenker, R and Ng, P. (2002).  SparseM:  A Sparse Matrix Package for \R,\cr
\url{http://www.econ.uiuc.edu/~roger/research}}
\seealso{
\code{SparseM.ops}, 
\code{SparseM.solve}, 
\code{SparseM.ontology} 
}
\examples{
a_rnorm(20*5)
A_matrix(a,20,5)
A[row(A)>col(A)+4|row(A)<col(A)+3] <- 0
b_rnorm(20*5)
B_matrix(b,20,5)
B[row(A)>col(A)+2|row(A)<col(A)+2] <- 0
image(as.matrix.csr(A)\%*\%as.matrix.csr(t(B)))
}
\keyword{hplot}
\keyword{algebra}
